requirement.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==22.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
chardet==4.0.0
charset-normalizer==2.1.1
Colorings==1.0.2
discord.py-self==1.9.2
discord.py-self-embed==1.0.6
idna==3.4
multidict==6.0.2
requests==2.28.1
soupsieve==2.3.2.post1
typing-extensions==4.4.0
urllib3==1.26.12
yarl==1.8.1
